# Optimize compact during framing