--- a/akka-stream/src/main/scala/akka/stream/scaladsl/Framing.scala
+++ b/akka-stream/src/main/scala/akka/stream/scaladsl/Framing.scala
@@ -5,16 +5,16 @@
 package akka.stream.scaladsl
 
 import java.nio.ByteOrder
-import java.util
 
 import akka.NotUsed
 import akka.stream.impl.Stages.DefaultAttributes
 import akka.stream.impl.fusing.GraphStages.SimpleLinearGraphStage
-import akka.stream.{ Attributes, FlowShape, Inlet, Outlet }
 import akka.stream.stage._
+import akka.stream.{ Attributes, FlowShape, Inlet, Outlet }
 import akka.util.{ ByteIterator, ByteString }
 
 import scala.annotation.tailrec
+import scala.reflect.ClassTag
 
 object Framing {
 
@@ -27,9 +27,9 @@ object Framing {
    * If there are buffered bytes (an incomplete frame) when the input stream finishes and ''allowTruncation'' is set to
    * false then this Flow will fail the stream reporting a truncated frame.
    *
-   * @param delimiter The byte sequence to be treated as the end of the frame.
-   * @param allowTruncation If `false`, then when the last frame being decoded contains no valid delimiter this Flow
-   *                        fails the stream instead of returning a truncated frame.
+   * @param delimiter          The byte sequence to be treated as the end of the frame.
+   * @param allowTruncation    If `false`, then when the last frame being decoded contains no valid delimiter this Flow
+   *                           fails the stream instead of returning a truncated frame.
    * @param maximumFrameLength The maximum length of allowed frames while decoding. If the maximum length is
    *                           exceeded this Flow will fail the stream.
    */
@@ -44,12 +44,12 @@ object Framing {
    * If the input stream finishes before the last frame has been fully decoded, this Flow will fail the stream reporting
    * a truncated frame.
    *
-   * @param fieldLength The length of the "size" field in bytes
-   * @param fieldOffset The offset of the field from the beginning of the frame in bytes
+   * @param fieldLength        The length of the "size" field in bytes
+   * @param fieldOffset        The offset of the field from the beginning of the frame in bytes
    * @param maximumFrameLength The maximum length of allowed frames while decoding. If the maximum length is exceeded
    *                           this Flow will fail the stream. This length *includes* the header (i.e the offset and
    *                           the length of the size field)
-   * @param byteOrder The ''ByteOrder'' to be used when decoding the field
+   * @param byteOrder          The ''ByteOrder'' to be used when decoding the field
    */
   def lengthField(
     fieldLength:        Int,
@@ -68,16 +68,16 @@ object Framing {
    * If the input stream finishes before the last frame has been fully decoded, this Flow will fail the stream reporting
    * a truncated frame.
    *
-   * @param fieldLength The length of the "size" field in bytes
-   * @param fieldOffset The offset of the field from the beginning of the frame in bytes
+   * @param fieldLength        The length of the "size" field in bytes
+   * @param fieldOffset        The offset of the field from the beginning of the frame in bytes
    * @param maximumFrameLength The maximum length of allowed frames while decoding. If the maximum length is exceeded
    *                           this Flow will fail the stream. This length *includes* the header (i.e the offset and
    *                           the length of the size field)
-   * @param byteOrder The ''ByteOrder'' to be used when decoding the field
-   * @param computeFrameSize This function can be supplied if frame size is varied or needs to be computed in a special fashion.
-   *                         For example, frame can have a shape like this: `[offset bytes][body size bytes][body bytes][footer bytes]`.
-   *                         Then computeFrameSize can be used to compute the frame size: `(offset bytes, computed size) => (actual frame size)`.
-   *                         ''Actual frame size'' must be equal or bigger than sum of `fieldOffset` and `fieldLength`, the operator fails otherwise.
+   * @param byteOrder          The ''ByteOrder'' to be used when decoding the field
+   * @param computeFrameSize   This function can be supplied if frame size is varied or needs to be computed in a special fashion.
+   *                           For example, frame can have a shape like this: `[offset bytes][body size bytes][body bytes][footer bytes]`.
+   *                           Then computeFrameSize can be used to compute the frame size: `(offset bytes, computed size) => (actual frame size)`.
+   *                           ''Actual frame size'' must be equal or bigger than sum of `fieldOffset` and `fieldLength`, the operator fails otherwise.
    *
    */
   def lengthField(
@@ -197,19 +197,20 @@ object Framing {
       private val firstSeparatorByte = separatorBytes.head
       private var buffer = ByteString.empty
       private var nextPossibleMatch = 0
+      private val indices = new LightArray[(Int, Int)](256)
 
       override def onPush(): Unit = {
         buffer ++= grab(in)
-        doParse()
+        searchIndices()
       }
 
-      override def onPull(): Unit = doParse()
+      override def onPull(): Unit = searchIndices()
 
       override def onUpstreamFinish(): Unit = {
         if (buffer.isEmpty) {
           completeStage()
         } else if (isAvailable(out)) {
-          doParse()
+          searchIndices()
         } // else swallow the termination and wait for pull
       }
 
@@ -224,66 +225,103 @@ object Framing {
         } else pull(in)
       }
 
-      private def doParse(): Unit = {
-        val possibleMatchPos = buffer.indexOf(firstSeparatorByte, from = nextPossibleMatch)
-        if (possibleMatchPos > maximumLineBytes) {
-          failStage(new FramingException(s"Read ${buffer.size} bytes " +
-            s"which is more than $maximumLineBytes without seeing a line terminator"))
-        } else if (possibleMatchPos == -1 && buffer.size > maximumLineBytes) {
-          failStage(new FramingException(s"Read ${buffer.size} bytes " +
-            s"which is more than $maximumLineBytes without seeing a line terminator"))
-        } else {
-          doSubParse(new util.ArrayDeque[ByteString](), possibleMatchPos)
-        }
-      }
-
       @tailrec
-      private def doSubParse(results: util.ArrayDeque[ByteString], pos: Int = -2): Unit = {
-        val possibleMatchPos = {
-          // Avoid to recompute when pos > -2 because we already compute a pos earlier.
-          if (pos > -2) pos
-          else buffer.indexOf(firstSeparatorByte, from = nextPossibleMatch)
-        }
+      private def searchIndices(): Unit = {
+        // Next possible position for the delimiter
+        val possibleMatchPos = buffer.indexOf(firstSeparatorByte, from = nextPossibleMatch)
 
-        // No match but we can release previous results if exists
-        if (possibleMatchPos == -1) {
-          nextPossibleMatch = buffer.size
-          if (results.isEmpty) {
+        // Retrive previous position
+        val previous = indices.lastOption.map(_._2 + separatorBytes.size).getOrElse(0)
+        if (possibleMatchPos - previous > maximumLineBytes) {
+          failStage(new FramingException(s"Read ${possibleMatchPos - previous} bytes " +
+            s"which is more than $maximumLineBytes without seeing a line terminator"))
+        } else if (possibleMatchPos == -1) {
+          if (buffer.size - previous > maximumLineBytes)
+            failStage(new FramingException(s"Read ${buffer.size - previous} bytes " +
+              s"which is more than $maximumLineBytes without seeing a line terminator"))
+          else {
             // No matching character, we need to accumulate more bytes into the buffer
-            tryPull()
-          } else {
-            buffer = buffer.compact
-            emitMultiple(out, results.iterator())
+            nextPossibleMatch = buffer.size
+            doParse()
           }
         } else if (possibleMatchPos + separatorBytes.size > buffer.size) {
           // We have found a possible match (we found the first character of the terminator
           // sequence) but we don't have yet enough bytes. We remember the position to
           // retry from next time.
           nextPossibleMatch = possibleMatchPos
-
-          if (results.isEmpty) {
-            // No matching character, we need to accumulate more bytes into the buffer
-            tryPull()
-          } else {
-            buffer = buffer.compact
-            emitMultiple(out, results.iterator())
-          }
+          doParse()
         } else if (buffer.slice(possibleMatchPos, possibleMatchPos + separatorBytes.size) == separatorBytes) {
-          // Found a match
-          results.add(buffer.slice(0, possibleMatchPos).compact)
-          buffer = buffer.drop(possibleMatchPos + separatorBytes.size)
-          nextPossibleMatch = 0
-          if (isClosed(in) && buffer.isEmpty) {
-            emitMultiple(out, results.iterator())
-            completeStage()
+          // Found a match, mark start and end position and iterate if possible
+          indices += (previous, possibleMatchPos)
+          nextPossibleMatch = possibleMatchPos + separatorBytes.size
+          if (nextPossibleMatch == buffer.size || indices.isFull) {
+            doParse()
           } else {
-            doSubParse(results)
+            searchIndices()
           }
         } else {
           // possibleMatchPos was not actually a match
           nextPossibleMatch += 1
-          doSubParse(results)
+          searchIndices()
         }
+      }
+
+      private def doParse(): Unit =
+        if (indices.isEmpty) tryPull()
+        else if (indices.length == 1) {
+          // Emit result and compact buffer
+          val indice = indices(0)
+          push(out, buffer.slice(indice._1, indice._2).compact)
+          reset()
+          if (isClosed(in) && buffer.isEmpty) completeStage()
+        } else {
+          // Emit results and compact buffer
+          emitMultiple(out, new FrameIterator(), () ⇒ {
+            reset()
+            if (isClosed(in) && buffer.isEmpty) completeStage()
+          })
+        }
+
+      private def reset(): Unit = {
+        buffer = buffer.drop(indices.lastOption.map(_._2 + separatorBytes.size).getOrElse(0)).compact
+        indices.setLength(0)
+        nextPossibleMatch = 0
+      }
+
+      // Iterator able to iterate over precompute frame based on start and end position
+      private class FrameIterator(private var index: Int = 0) extends Iterator[ByteString] {
+        def hasNext: Boolean = index != indices.length
+
+        def next(): ByteString = {
+          val indice = indices(index)
+          index += 1
+          buffer.slice(indice._1, indice._2).compact
+        }
+      }
+
+      // Basic array implementation that allow unsafe resize.
+      private class LightArray[T: ClassTag](private val capacity: Int, private var index: Int = 0) {
+
+        private val underlying = Array.ofDim[T](capacity)
+
+        def apply(i: Int) = underlying(i)
+
+        def +=(el: T): Unit = {
+          underlying(index) = el
+          index += 1
+        }
+
+        def isEmpty: Boolean = length == 0
+
+        def isFull: Boolean = capacity == length
+
+        def setLength(length: Int): Unit = index = length
+
+        def length: Int = index
+
+        def lastOption: Option[T] =
+          if (index > 0) Some(underlying(index - 1))
+          else None
       }
       setHandlers(in, out, this)
     }