Currently `Framing.delimiter` can pass a significant amount of time allocating byte arrays instead of parsing on a high load.
This is caused by the use of the method `compact` each time we found a frame.
In some case, we can get multiples frames in one `ByteString` [[Ref](https://github.com/amannocci/streamy/issues/38)] and in these cases compacting every time is inefficient.
A solution can be to parse every frame and then compact to let GC do is work.
I have done a small improvement and a JMH to test this solution.
I have obtained some interesting results:

Master branch:
```bash
[info] Benchmark                   (framePerSeq)   Mode  Cnt      Score      Error  Units
[info] FramingBenchmark.framing               32  thrpt    5  93675.026 ± 1381.528  ops/s
[info] FramingBenchmark.framing               64  thrpt    5  34122.613 ±  540.357  ops/s
[info] FramingBenchmark.framing              128  thrpt    5  11310.841 ±  146.387  ops/s
```

PR branch:
```bash
[info] Benchmark                   (framePerSeq)   Mode  Cnt       Score      Error  Units
[info] FramingBenchmark.framing               32  thrpt    5  232453.491 ± 1542.536  ops/s
[info] FramingBenchmark.framing               64  thrpt    5  119805.343 ± 1235.092  ops/s
[info] FramingBenchmark.framing              128  thrpt    5   57768.423 ±  448.558  ops/s
```

Tell me what you thing about it.