@raboof 2018-11-12
@amannocci that looks very interesting!

Jenkins fails because it found some formatting inconsistencies, could you run `sbt scalariformFormat` on this branch to autocorrect this?