@amannocci 2018-11-14
I will do a more complete bench scenario to avoid performance regression in other scenario.