@amannocci 2018-11-15
I change the bench to cover any real world case.
Here the result of the bench on master and optimize-framing branch.
Clearly there is something to do to avoid extra overhead in simple framing case.

Master branch:
```bash
[info] Benchmark                                    (framePerSeq)  (messageSize)   Mode  Cnt        Score         Error  Units
[info] FramingBenchmark.framing                                 1             32  thrpt    3  7221077.788 ± 1139069.107  ops/s
[info] FramingBenchmark.framing                                 1             64  thrpt    3  6398577.765 ±   82217.022  ops/s
[info] FramingBenchmark.framing                                 1            128  thrpt    3  4986464.141 ±  530256.354  ops/s
[info] FramingBenchmark.framing                                 1            256  thrpt    3  3596370.029 ±  239134.562  ops/s
[info] FramingBenchmark.framing                                 1            512  thrpt    3  2427753.389 ±   24783.443  ops/s
[info] FramingBenchmark.framing                                 1           1024  thrpt    3  1268860.506 ±   67526.039  ops/s
[info] FramingBenchmark.framing                                 8             32  thrpt    3  1130189.150 ±  121845.454  ops/s
[info] FramingBenchmark.framing                                 8             64  thrpt    3   914422.925 ±   81403.339  ops/s
[info] FramingBenchmark.framing                                 8            128  thrpt    3   645391.687 ±   47992.862  ops/s
[info] FramingBenchmark.framing                                 8            256  thrpt    3   379470.335 ±    9073.314  ops/s
[info] FramingBenchmark.framing                                 8            512  thrpt    3   217213.167 ±    7767.610  ops/s
[info] FramingBenchmark.framing                                 8           1024  thrpt    3   121411.683 ±    2793.492  ops/s
[info] FramingBenchmark.framing                                16             32  thrpt    3   570088.833 ±   22984.600  ops/s
[info] FramingBenchmark.framing                                16             64  thrpt    3   416057.197 ±   26740.257  ops/s
[info] FramingBenchmark.framing                                16            128  thrpt    3   249721.777 ±    3544.939  ops/s
[info] FramingBenchmark.framing                                16            256  thrpt    3   150212.437 ±    8323.327  ops/s
[info] FramingBenchmark.framing                                16            512  thrpt    3    89136.428 ±   21242.046  ops/s
[info] FramingBenchmark.framing                                16           1024  thrpt    3    49426.709 ±     217.297  ops/s
[info] FramingBenchmark.framing                                32             32  thrpt    3   242966.718 ±    7736.077  ops/s
[info] FramingBenchmark.framing                                32             64  thrpt    3   152051.220 ±    6635.331  ops/s
[info] FramingBenchmark.framing                                32            128  thrpt    3    92303.483 ±    3484.725  ops/s
[info] FramingBenchmark.framing                                32            256  thrpt    3    57754.352 ±    1588.937  ops/s
[info] FramingBenchmark.framing                                32            512  thrpt    3    33701.421 ±    1963.857  ops/s
[info] FramingBenchmark.framing                                32           1024  thrpt    3    17597.413 ±    1307.124  ops/s
[info] FramingBenchmark.framing                                64             32  thrpt    3    84215.651 ±    2362.604  ops/s
[info] FramingBenchmark.framing                                64             64  thrpt    3    54211.251 ±    2880.419  ops/s
[info] FramingBenchmark.framing                                64            128  thrpt    3    34867.730 ±    2840.094  ops/s
[info] FramingBenchmark.framing                                64            256  thrpt    3    20457.160 ±     454.228  ops/s
[info] FramingBenchmark.framing                                64            512  thrpt    3    10832.474 ±    1417.893  ops/s
[info] FramingBenchmark.framing                                64           1024  thrpt    3     3758.286 ±     341.820  ops/s
[info] FramingBenchmark.framing                               128             32  thrpt    3    28331.829 ±    1813.093  ops/s
[info] FramingBenchmark.framing                               128             64  thrpt    3    18936.636 ±    1216.616  ops/s
[info] FramingBenchmark.framing                               128            128  thrpt    3    11401.925 ±     650.561  ops/s
[info] FramingBenchmark.framing                               128            256  thrpt    3     5911.314 ±     404.484  ops/s
[info] FramingBenchmark.framing                               128            512  thrpt    3     3033.682 ±     475.794  ops/s
[info] FramingBenchmark.framing                               128           1024  thrpt    3     1522.262 ±      61.807  ops/s
```

PR branch:
```bash
[info] Benchmark                                    (framePerSeq)  (messageSize)   Mode  Cnt        Score        Error  Units
[info] FramingBenchmark.framing                                 1             32  thrpt    3  4140643.756 ± 117994.325  ops/s
[info] FramingBenchmark.framing                                 1             64  thrpt    3  3445035.336 ± 135744.701  ops/s
[info] FramingBenchmark.framing                                 1            128  thrpt    3  2281914.829 ± 214024.040  ops/s
[info] FramingBenchmark.framing                                 1            256  thrpt    3  1843715.219 ± 119141.226  ops/s
[info] FramingBenchmark.framing                                 1            512  thrpt    3  1129502.831 ±  43045.165  ops/s
[info] FramingBenchmark.framing                                 1           1024  thrpt    3   840100.123 ±  36372.909  ops/s
[info] FramingBenchmark.framing                                 8             32  thrpt    3  1241650.254 ±  67432.715  ops/s
[info] FramingBenchmark.framing                                 8             64  thrpt    3  1049905.607 ±  69195.227  ops/s
[info] FramingBenchmark.framing                                 8            128  thrpt    3   779239.109 ±  31026.838  ops/s
[info] FramingBenchmark.framing                                 8            256  thrpt    3   538565.208 ±  15330.974  ops/s
[info] FramingBenchmark.framing                                 8            512  thrpt    3   326470.904 ±   6573.503  ops/s
[info] FramingBenchmark.framing                                 8           1024  thrpt    3   184835.847 ±   2894.042  ops/s
[info] FramingBenchmark.framing                                16             32  thrpt    3   669010.159 ±  27369.045  ops/s
[info] FramingBenchmark.framing                                16             64  thrpt    3   566447.225 ±   4240.321  ops/s
[info] FramingBenchmark.framing                                16            128  thrpt    3   406149.526 ±  10544.927  ops/s
[info] FramingBenchmark.framing                                16            256  thrpt    3   282131.250 ±  13026.043  ops/s
[info] FramingBenchmark.framing                                16            512  thrpt    3   173743.669 ±   5724.513  ops/s
[info] FramingBenchmark.framing                                16           1024  thrpt    3    95083.564 ±   3022.305  ops/s
[info] FramingBenchmark.framing                                32             32  thrpt    3   359371.234 ±  36495.590  ops/s
[info] FramingBenchmark.framing                                32             64  thrpt    3   303963.585 ±   5992.681  ops/s
[info] FramingBenchmark.framing                                32            128  thrpt    3   222006.841 ±   7344.223  ops/s
[info] FramingBenchmark.framing                                32            256  thrpt    3   148895.455 ±   3413.631  ops/s
[info] FramingBenchmark.framing                                32            512  thrpt    3    99121.709 ±   3286.699  ops/s
[info] FramingBenchmark.framing                                32           1024  thrpt    3    52000.776 ±   4619.679  ops/s
[info] FramingBenchmark.framing                                64             32  thrpt    3   188142.065 ±   9357.309  ops/s
[info] FramingBenchmark.framing                                64             64  thrpt    3   156859.033 ±   1987.992  ops/s
[info] FramingBenchmark.framing                                64            128  thrpt    3   118968.047 ±   2929.047  ops/s
[info] FramingBenchmark.framing                                64            256  thrpt    3    82085.249 ±   5358.742  ops/s
[info] FramingBenchmark.framing                                64            512  thrpt    3    49112.780 ±   2202.507  ops/s
[info] FramingBenchmark.framing                                64           1024  thrpt    3    26572.792 ±   1425.715  ops/s
[info] FramingBenchmark.framing                               128             32  thrpt    3    97129.392 ±    892.098  ops/s
[info] FramingBenchmark.framing                               128             64  thrpt    3    81033.479 ±   4434.710  ops/s
[info] FramingBenchmark.framing                               128            128  thrpt    3    58550.705 ±   1728.965  ops/s
[info] FramingBenchmark.framing                               128            256  thrpt    3    38378.448 ±   1109.805  ops/s
[info] FramingBenchmark.framing                               128            512  thrpt    3    23126.931 ±   1032.074  ops/s
[info] FramingBenchmark.framing                               128           1024  thrpt    3    12442.198 ±    381.179  ops/s
```

I think we can perform better in every case.
I will try with another approach in three phases.
* First we can try to search every delimiter in the current frame 
* Then we split every frame using slice and compact based on the previous search.
* And finally we can compact the whole buffer

It will avoid any unneeded allocation and we can deal with single frame and multiples frames easily by allocating a collection with the right size or by just passing the element.
Tell me what you think about this solution ?